feat: training.py now calls NumberDrawing.py
After training.py finishes training the dataset,
it will now bring up the pygame where the user
can draw a number.

NumberDrawing.py was made into a function so it could
be called by other files.  Next goal is to make
NumberDrawing.py return the image data of the game
so training can evaluate it.